[gdef] Move glyph_props around
