+++ Cargo.toml
@@ -14,7 +14,7 @@ exclude = ["benches/", "tests/"]
 bitflags = "2.9"
 bytemuck = { version = "1.22", features = ["extern_crate_alloc"] }
 core_maths = "0.1" # only for no_std builds
-read-fonts = { version = "0.31.3", default-features = false, features = ["libm"] }
+read-fonts = { path = "../fontations/read-fonts", default-features = false, features = ["libm"] }
 smallvec = "1.14"
 
 [features]
