+++ src/hb/ot_layout.rs
@@ -85,7 +85,7 @@ pub fn _hb_ot_layout_set_glyph_props(face: &hb_font_t, buffer: &mut hb_buffer_t)
 
     let len = buffer.len;
     for info in &mut buffer.info[..len] {
-        info.set_glyph_props(face.glyph_props(info.as_glyph()));
+        info.set_glyph_props(face.ot_tables.glyph_props(info.as_glyph()));
         info.set_lig_props(0);
     }
 }
